[bug-fix]: make account page responsive on smaller devices
This commit brings a patch to make the accounts page look better
on smaller devices. This is by no means a proper fix. Ideally I'd
extract the components freemius sdk uses and make them responsive
with respective CSS and discard all ID based CSS codes. But that
is a task for another day. With this patch, the accounts page will
look good as long as there's no change in the HTML markup.

Fixes #453

Also squashing all 6 commits together so that the history remains
clean, but leaving the commit messages below for the record.

This is the 1st commit message:

[bug-fix]: make account page look better on mobile devices

I could not figure out the scss build process. So I just patched the css
file directly. Treat this as W.I.P. As soon as I figure out the scss
build, I will edit the relevant scss files.

This is the commit message #2:

[bug-fix]: fix accounts page mobile responsive view

Do the changes in the assets/scss/admin/account.scss file and build
it with globally installed sass. The next step would be to fix the
build scripts and have some sort of automation to make these tasks
easier in future.

Fixes #453

This is the commit message #3:

[fix]: use standard screen size in media query

Instead of the 520px, use a standard size 640px in the media query.
Also Make the tbody display as block too to take up the full
available width.

This is the commit message #4:

[fix]: value of max-width in accounts css

Since we are doing max-width, instead of mobile first
min-width, the value should be one less than the
standard size.

This is the commit message #5:

[revert]: discard changes to source css files

This prevents conflict on merging onto develop branch.

This is the commit message #6:

[fix]: format comments of source scss file per convention